Examples split up in Version_01 and Version_02
Version_02 added. this example contains an adafruit_displayio.flipclock with sercom_I2C connection. The device that performs the role of Main requests the device that performs the role of Sensor for a datetime. The received datetime message from the device with the role Sensor is then prepared (hh and mm extracted and feeded to the flipclock.